+++ test/coverage.test.js
@@ -18,12 +18,28 @@ var verifyCoverage = function *(file, t) {
 var sdps = [
   'normal.sdp',
   'hacky.sdp',
+  'icelite.sdp',
   'jssip.sdp',
   'jsep.sdp',
-  //'alac.sdp', // deliberate invalids
-  //'onvif.sdp', // SHOULD PASS
+  // 'alac.sdp', // deliberate invalids
+  'onvif.sdp',
   'ssrc.sdp',
-  'simulcast.sdp'
+  'simulcast.sdp',
+  'st2022-6.sdp',
+  // 'st2110-20.sdp', // deliberate invalids
+  'sctp-dtls-26.sdp',
+  'extmap-encrypt.sdp',
+  'dante-aes67.sdp',
+  'bfcp.sdp',
+  'tcp-active.sdp',
+  'tcp-passive.sdp',
+  'mediaclk-avbtp.sdp',
+  'mediaclk-ptp-v2-w-rate.sdp',
+  'mediaclk-ptp-v2.sdp',
+  'mediaclk-rtp.sdp',
+  'ts-refclk-media.sdp',
+  'ts-refclk-sess.sdp',
+  'rtcp-fb.sdp'
 ];
 
 sdps.forEach((name) => {